Create README
This project is provided for informational and educational purposes. It is distributed without warranties of any kind, express or implied, including, but not limited to, warranties of merchantability or fitness for a particular purpose. By using this project, you agree to do so at your own risk and not to hold contributors responsible for any damage, loss of data, or other harm resulting from the use of this code.